LogicalOp
So Using logical operations is like using the 'if/else' statements. But in truth there is a lot more to be added. You can use a public bool to have a true or false situation. After having the if else function written out, you must have one more else to print a phrase/ sentence when the if else statement requirements cannot be met.